# Major hill shading performance improvements