* Major performance improvements when using hill shading.
   - Improved hill shaded tiles overlapping.

       Overlapping is now only done when necessary during drawing, not during loading when there is not enough information about the required tiles. As an (worst case scenario) example, imagine we want to display a map tile where 4 hill-shaded tiles meet on large zoom level. The previous algorithm would wastefully load and process 12 shading tiles just to draw this one map tile: 4 required center tiles, but then also 8 "invisible" tiles: 2 tiles north, 2 tiles west, 2 tiles south, and 2 tiles east of the 4 center tiles, 12 tiles total. Now only the 4 center tiles are loaded and overlapped, since that is all that is needed to render the map tile.

   - Performance of the high resolution (i.e. bicubic) hill shading was improved by almost 4x (four times faster).

       Quick benchmarks show the following times to process a single 1" HGT file (that is 25MB or almost 13M points per file):

       - 900 ms (new high resolution algorithm)
       - 550 ms (new standard resolution algorithm)
       
         Comparing this to a previous version:

       - 3200 ms (v0.22.0 high resolution algorithm)
       - 1550 ms (v0.22.0 diffuse algorithm)
       
       Total times are much more drastic. These are times until the map is completely loaded (DEM files were processed one by one):
       
       - 3.9 seconds (new high resolution algorithm, loading 4 tiles)
       - 2.5 seconds (new standard resolution algorithm, loading 4 tiles)
       - 40 seconds (v0.22.0 high resolution algorithm, loading 12 tiles)
       - 19 seconds (v0.22.0 diffuse algorithm, loading 12 tiles)
      
      All algorithms were on default settings (including preprocessing enabled), on an Android device with 4x 2.5GHz cores and 4x 1.9GHz cores.

* Major visual improvements when using hill shading. Removed nearly half a dozen causes of various artifacts on hill shaded tiles. This also resulted in elegant code simplifications and performance improvements.

   - Artifacts due to missing input data and poor preprocessing
   - Artifacts due to hill shading bitmap overlapping misalignments
   - Artifacts due to hill shading bitmap filtering
   - Artifacts due to poor thread synchronization while drawing hill shading bitmaps

* Input data preprocessing is now a separate step, and an option (enabled by default). Preprocessing removes invalid values from the input data by filling some more clever value, like an average of neighboring points. The preprocessing algorithm is now smarter than before, previously it just filled a fixed neighboring value without checking that it also might be invalid. It is now also being done during processing phase, not reading phase, i.e. on a computing thread pool (thus multi-threaded).

* Major optimization and performance improvements to AwtCanvas.shadeBitmap().
   
   - Preventing needless up-scaling of the entire hill shading bitmap, that previously wasted large amounts of memory and CPU time when only a small part of the bitmap is really needed. This was especially prominent on larger zoom levels, and could potentially cause an OOM exception.


Examples of artifacts that were successfully handled:

![2024-10_artifact1r](https://github.com/user-attachments/assets/7a993a46-e50f-4d6a-bfa3-089540dae970) ![2024-10_artifact2r](https://github.com/user-attachments/assets/8d37ed5c-2fc8-4479-b8fd-f87cd21a09f4) ![2024-10_artifact3r](https://github.com/user-attachments/assets/6c313fd6-f946-4ae8-af5a-f51b43ef82ca)